ci-cd-fargate/flask-docker-app/requirements.txt

Flask==2.2.2
flask-cors
watchtower
mysql-connector-python